requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
python-dotenv
python-multipart
jinja2
pydantic
Pillow
